[SplitModule] Move externalize helpers to SplitModuleCommon
Create a dedicated home for utilities shared by the module splitting
implementations, and rename __llvmsplit_unnamed to
__llvmsplit_externalize_unnamed to reflect where the name comes from.
